Allow multiple RN instances to run at the same time
Summary:
Users can instantiate multiple instances of React Native in their apps. Technically, different Activity could potentially have different React Native instances.
However, currently, if you try to run multiple instances, the app will crash with the message:

```
terminating due to uncaught exception of type std::runtime_error: Feature flags cannot be overridden more than once
```

This happens also when the feature flags we would like to set are the same that we already applied. This should be an allowed scenario because reapplying the excatly same features flags should have no effect on React native and that's not the use case we want to forbid.

With this change, we are creating a static variable that checks whether we already apply that set of feature flags and it allows you to create multiple instances by keeping the same flags

## Changelog:
[iOS][Fixed] - Allow multiple RN instances to run at the same time

Differential Revision: D69402863