Move RuntimeExecutor up
Summary:
# Changelog: [Internal]

We need an ability to report the JavaScript thread id to Performance Tracer.

My thinking is that InstanceAgent should be responsible for it, since React Native Instance coordinates the threads: Main vs JavaScript.

When tracing will start `InstanceTarget` will register itself for tracing and will execute the code on JavaScript thread to record it.

Differential Revision: D69530983