Generalize native module registration to nativeModules (#49653)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/49653

This change generalizes the C++ Module registration approach to register also native modules through Codegen.

This allow us to fully remove macros like `RCT_EXPORT_MODULE`.

## Problem
As of today, it is not possible to create a pure C++ TM and to register it through a Swift AppDelegate

## Solution
We can create a pod that can be imported in a Swift AppDelegate and that offer some pure Objective-C classes.

These classes contains a provider that can be instantiated in Swift.

The TurboModule manager delegate will ask the AppDelegate about the presence of some provider that can instantiate a pure C++ turbomodule with a given name.

The provider has an empty interface, but the implementation contains a function that can actually instantiate the TM. The function is implemented in an Objective-C++ class that imports the pure C++ turbomodule and creates it.

The TMManager extends the provider through a category to attaach the signature of the function that is implemented by the provider.

The last diff in this stack contains an exaple on how to implement this.

## Changelog:
[iOS][Added] - Generalize the mechanism to register Cxx turbomodules to NativeModules

Differential Revision: D70119870