Move environment types (globals) from flow-typed/environment to src/types
Summary:
Changelog: [internal]

This moves the type definitions for the globals defined by React Native from the repository root to the `react-native` package, in the same directory as the existing TypeScript definitions. This will make it easier for end users to configure the globals from RN using the right source of truth.

It also removes the type definitions for service workers, as it's unused in RN.

Differential Revision: D70322032