Updated folly to use folly-config.h file
Due to the way folly is set up, it needs configuration to be set when building it and also when you reference it through header files. This has led to the folly config having to be set in almost all our podspec files - and also in a lot of OSS libraries.

This configuration change makes use of the option in folly to configure using a header file - which will be used both when building and also when referencing it from other libraries - without having to pass folly compiler flags in all podspec files.

This is now solved by:

1. In prebuild's setup step (prepareScript) we create the folly-config.h file with the necessary configuration
2. This file is included in the distributed headers

Whenever anyone references folly they also include this file - and no extra config flags are needed.