Disable no-react-native-imports rule in Fantom tests (#49911)
Summary:

Changelog: [internal]

The approach in Fantom tests is to use the public API as much as we can, but forcing us to use relative imports in tests makes it harder to see what's the public API and what's not. This disables the lint rule so we can use package imports in Fantom tests without warnings/errors.

Reviewed By: sammy-SC

Differential Revision: D70779721